requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
